requirements.txt
sseclient
django>=2.0
django-bulk-update
django-cache-machine
psycopg2
django-rest-framework
python-dateutil
